+++ GiocoSPACCA/src/application/Player/PlayerList.java
@@ -2,6 +2,44 @@ package application.Player;
 
 public class PlayerList {
 
+	private class Node {
+
+		private Player current;
+		private Player next;
+		private Player previous;
+		
+		public Node(Player pre, Player cur, Player next) {
+			this.previous = pre;
+			this.current = cur;
+			this.next = next;
+		}
+		
+		public Node(Player cur, Player next) { //nel caso di due giocatori
+			this.current = cur;
+			this.next = next;
+		}
+		
+		public Player getPrevious() {
+			return this.previous;
+		}
+		
+		public Player getCurrent() {
+			return this.current;
+		}
+		
+		public Player getNext() {
+			return this.next;
+		}
+		
+		public void setPrevius(Player p) {
+			this.previous = p;
+		}
+		
+		public void setNext(Player p) {
+			this.next=p;
+		}
+	}
+
 	private Node[] list;
 	
 	public PlayerList(Player p1, Player p2) {
